refactor : "box collder"

+++ TDS_Project/Assets/2. Scripts/MonsterData.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 public class MonsterData : ScriptableObject
 {
     [SerializeField]
-    private float damage;
+    private int damage;
 
-    public float Damage // 몬스터가 주는 데미지는 고정값으로 지정하며, 모든 몬스터의 데미지값은 동일하다.
+    public int Damage // 몬스터가 주는 데미지는 고정값으로 지정하며, 모든 몬스터의 데미지값은 동일하다.
     {
         get
         {
